rename homepage to testingpage
