+++ ethernet-view/src/pages/LoggerPage/LoggerPage.tsx
@@ -1,0 +1,6 @@
+
+export const LoggerPage = () => {
+  return (
+    <div>LoggerPage</div>
+  )
+}
