+++ ethernet-view/src/pages/TestingPage/ChartsColumn/ChartsColumn.tsx
