+++ ethernet-view/src/pages/TestingPage/ReceiveColumn/ReceiveColumn.module.scss
