+++ ethernet-view/src/pages/TestingPage/ReceiveColumn/ReceiveColumn.tsx
