+++ ethernet-view/src/pages/TestingPage/TestingPage.tsx
@@ -1,16 +1,16 @@
 import { SplitLayout } from "layouts/SplitLayout/SplitLayout";
 import { Orientation } from "hooks/useSplit/Orientation";
-import { ReceiveColumn } from "pages/HomePage/ReceiveColumn/ReceiveColumn";
-import { OrderColumn } from "pages/HomePage/OrderColumn/OrderColumn";
-import { MessagesColumn } from "pages/HomePage/MessagesColumn/MessagesColumn";
+import { ReceiveColumn } from "pages/TestingPage/ReceiveColumn/ReceiveColumn";
+import { OrderColumn } from "pages/TestingPage/OrderColumn/OrderColumn";
+import { MessagesColumn } from "pages/TestingPage/MessagesColumn/MessagesColumn";
 import { ChartsColumn } from "./ChartsColumn/ChartsColumn";
-import styles from "pages/HomePage/HomePage.module.scss";
+import styles from "pages/TestingPage/TestingPage.module.scss";
 import incomingMessage from "assets/svg/incoming-message.svg"
 import paperAirplane from "assets/svg/paper-airplane.svg"
 import outgoingMessage from "assets/svg/outgoing-message.svg"
 import chart from "assets/svg/chart.svg"
 
-export const HomePage = () => {
+export const TestingPage = () => {
     return (
         <div id={styles.wrapper}>
             <div id={styles.body}>
